A bit of parsing cleanup
