+++ crates/rune/src/ast/expr_if.rs
@@ -1,4 +1,4 @@
-use crate::ast::{Attribute, Condition, Else, ExprBlock, ExprElse, ExprElseIf, If};
+use crate::ast;
 use crate::{Parse, ParseError, Parser, Spanned, ToTokens};
 
 /// An if statement: `if cond { true } else { false }`
@@ -18,13 +18,13 @@ use crate::{Parse, ParseError, Parser, Spanned, ToTokens};
 pub struct ExprIf {
     /// The `attributes` of the if statement
     #[rune(iter)]
-    pub attributes: Vec<Attribute>,
+    pub attributes: Vec<ast::Attribute>,
     /// The `if` token.
-    pub if_: If,
+    pub if_: ast::If,
     /// The condition to the if statement.
-    pub condition: Condition,
+    pub condition: ast::Condition,
     /// The body of the if statement.
-    pub block: Box<ExprBlock>,
+    pub block: Box<ast::Block>,
     /// Else if branches.
     #[rune(iter)]
     pub expr_else_ifs: Vec<ExprElseIf>,
@@ -37,16 +37,16 @@ impl ExprIf {
     /// Parse an if statement attaching the given attributes
     pub fn parse_with_attributes(
         parser: &mut Parser,
-        attributes: Vec<Attribute>,
+        attributes: Vec<ast::Attribute>,
     ) -> Result<Self, ParseError> {
         let if_ = parser.parse()?;
         let condition = parser.parse()?;
-        let block = Box::new(parser.parse()?);
+        let block = parser.parse()?;
         let mut expr_else_ifs = Vec::new();
         let mut expr_else = None;
 
-        while parser.peek::<Else>()? {
-            if parser.peek2::<If>()? {
+        while parser.peek::<ast::Else>()? {
+            if parser.peek2::<ast::If>()? {
                 expr_else_ifs.push(parser.parse()?);
                 continue;
             }
@@ -66,8 +66,30 @@ impl ExprIf {
 }
 
 impl Parse for ExprIf {
-    fn parse(parser: &mut Parser) -> Result<Self, ParseError> {
+    fn parse(parser: &mut Parser<'_>) -> Result<Self, ParseError> {
         let attributes = parser.parse()?;
         Self::parse_with_attributes(parser, attributes)
     }
+}
+
+/// An else branch of an if expression.
+#[derive(Debug, Clone, PartialEq, Eq, ToTokens, Parse, Spanned)]
+pub struct ExprElseIf {
+    /// The `else` token.
+    pub else_: ast::Else,
+    /// The `if` token.
+    pub if_: ast::If,
+    /// The condition for the branch.
+    pub condition: ast::Condition,
+    /// The body of the else statement.
+    pub block: Box<ast::Block>,
+}
+
+/// An else branch of an if expression.
+#[derive(Debug, Clone, PartialEq, Eq, ToTokens, Parse, Spanned)]
+pub struct ExprElse {
+    /// The `else` token.
+    pub else_: ast::Else,
+    /// The body of the else statement.
+    pub block: Box<ast::ExprBlock>,
 }
