+++ crates/rune/src/ast/mod.rs
@@ -27,8 +27,6 @@ mod expr_block;
 mod expr_break;
 mod expr_call;
 mod expr_closure;
-mod expr_else;
-mod expr_else_if;
 mod expr_field_access;
 mod expr_for;
 mod expr_group;
@@ -95,12 +93,10 @@ pub use self::expr_block::ExprBlock;
 pub use self::expr_break::{ExprBreak, ExprBreakValue};
 pub use self::expr_call::ExprCall;
 pub use self::expr_closure::ExprClosure;
-pub use self::expr_else::ExprElse;
-pub use self::expr_else_if::ExprElseIf;
 pub use self::expr_field_access::{ExprField, ExprFieldAccess};
 pub use self::expr_for::ExprFor;
 pub use self::expr_group::ExprGroup;
-pub use self::expr_if::ExprIf;
+pub use self::expr_if::{ExprElse, ExprElseIf, ExprIf};
 pub use self::expr_index_get::ExprIndexGet;
 pub use self::expr_index_set::ExprIndexSet;
 pub use self::expr_is::ExprIs;
